Finally fixed that fuckin issue goddamn
the key is really in using logging as much as possible. Just sit down for a while, and put logs everywhere to narrow in on the source of the problem. In the end, I had just screwed up a bunch of foreach and if loops, and it was getting confusing. But once I really figured out nothing else was the issue, I was able to wrap my head around the loops, and realize that it was missing a "return;" in the nested loop.